Replaced the old trainer with lightning
